# openapi.json
{"openapi":"3.1.0","info":{"title":"MarketTwin AI","description":"MarketTwin AI - Real-Time Omnichannel Customer Intelligence, Digital Twin, NBA Decisioning, and Consent-Aware Activation Platform","version":"2.0.0"},"paths":{"/api/health":{"get":{"tags":["Health"],"summary":"Health Check","description":"Inspects application status and database session connection viability.","operationId":"health_check_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/event":{"post":{"tags":["Events"],"summary":"Ingest Event","description":"Primary ingestion endpoint. Processes incoming customer actions, merges identities,\ntriggers behavioral updates, and outputs the updated digital twin.","operationId":"ingest_event_api_event_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Raw Event"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventIngestionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/customers":{"get":{"tags":["Customers"],"summary":"Get All Customers","description":"Retrieves all customer profiles stored in the system.","operationId":"get_all_customers_api_customers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CustomerProfileResponse"},"type":"array","title":"Response Get All Customers Api Customers Get"}}}}}}},"/api/customer/{customer_id}":{"get":{"tags":["Customers"],"summary":"Get Customer Profile","description":"Retrieves demographics and features associated with a customer ID.","operationId":"get_customer_profile_api_customer__customer_id__get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomerProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/customer/{customer_id}/twin":{"get":{"tags":["Customers"],"summary":"Get Customer Twin","description":"Retrieves the active behavior digital twin state for a customer ID.","operationId":"get_customer_twin_api_customer__customer_id__twin_get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TwinStateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/customer/{customer_id}/events":{"get":{"tags":["Customers"],"summary":"Get Customer Events","description":"Retrieves a list of all historical events associated with a customer ID.","operationId":"get_customer_events_api_customer__customer_id__events_get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventResponse"},"title":"Response Get Customer Events Api Customer  Customer Id  Events Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/nba/{customer_id}":{"get":{"tags":["Next Best Action","Next Best Action"],"summary":"Get next best action for a customer","description":"Returns the consent-gated next best action for the given customer.\n\nThe action has already been filtered through the consent engine —\nif the preferred channel is blocked, the response reflects the\nfallback action (or do_nothing if all channels are blocked).","operationId":"get_nba_api_nba__customer_id__get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NBAResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/nba/{customer_id}/explain":{"get":{"tags":["Next Best Action","Next Best Action"],"summary":"Verbose NBA explanation for ops/audit dashboards","description":"Returns the full NBA explanation including input snapshot and\nthreshold values used in the decision.  Intended for internal\ndashboards and model audits, not customer-facing APIs.","operationId":"explain_nba_api_nba__customer_id__explain_get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Explain Nba Api Nba  Customer Id  Explain Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/nba/{customer_id}/event":{"post":{"tags":["Next Best Action","Next Best Action"],"summary":"Process an event and return the updated NBA decision","description":"Runs the full pipeline:\n  Event → Twin Update → NBA Engine → Consent Gate → Final Action\n\nUse this endpoint to push raw behavioural events and immediately\nreceive the updated recommended action.","operationId":"process_event_api_nba__customer_id__event_post","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventPayload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Process Event Api Nba  Customer Id  Event Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/consent/{customer_id}":{"get":{"tags":["Consent","Consent"],"summary":"Retrieve the consent vault for a customer","description":"Returns the current per-channel consent state for a customer.\n\nIf the customer has no existing record, a default vault with all\nchannels enabled is created and returned (DPDP § opt-in at registration).","operationId":"get_consent_vault_api_consent__customer_id__get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsentVault"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Consent","Consent"],"summary":"Update consent preferences for a customer","description":"Partially updates the consent vault.  Only fields present in the\nrequest body are changed; omitted fields retain their current value.\n\nExample — disable WhatsApp only:\n```json\n{ \"whatsapp\": false }\n```\nThe response contains the full updated vault with a refreshed\nlast_updated timestamp.","operationId":"update_consent_api_consent__customer_id__post","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsentUpdateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsentVault"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/consent/{customer_id}/channel/{channel}":{"get":{"tags":["Consent","Consent"],"summary":"Check if a specific channel is consented for a customer","description":"Lightweight consent gate check — useful for downstream services\nthat need to verify a single channel without fetching the full vault.","operationId":"check_channel_consent_api_consent__customer_id__channel__channel__get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}},{"name":"channel","in":"path","required":true,"schema":{"type":"string","title":"Channel"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Check Channel Consent Api Consent  Customer Id  Channel  Channel  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/creative/generate":{"post":{"tags":["Creative Generation"],"summary":"Generate Creative","operationId":"generate_creative_api_creative_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreativeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/predictive/{customer_id}":{"get":{"tags":["Predictive Twin"],"summary":"Get Predictive Twin","operationId":"get_predictive_twin_api_predictive__customer_id__get","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","title":"Customer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/audience/segments":{"get":{"tags":["Audience Intelligence"],"summary":"Get Segments","operationId":"get_segments_api_audience_segments_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/audience/fatigue-heatmap":{"get":{"tags":["Audience Intelligence"],"summary":"Get Fatigue Heatmap","operationId":"get_fatigue_heatmap_api_audience_fatigue_heatmap_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/metrics/summary":{"get":{"tags":["Measurement Metrics"],"summary":"Get Summary","operationId":"get_summary_api_metrics_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/metrics/cohort-drift":{"get":{"tags":["Measurement Metrics"],"summary":"Get Cohort Drift","operationId":"get_cohort_drift_api_metrics_cohort_drift_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Read Root","description":"Standard landing route directing clients to documentation and health check.","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ConsentUpdateRequest":{"properties":{"email":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Email"},"whatsapp":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Whatsapp"},"push":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Push"}},"type":"object","title":"ConsentUpdateRequest","example":{"email":true,"push":true,"whatsapp":false}},"ConsentVault":{"properties":{"email":{"type":"boolean","title":"Email","default":true},"whatsapp":{"type":"boolean","title":"Whatsapp","default":true},"push":{"type":"boolean","title":"Push","default":true},"last_updated":{"type":"string","title":"Last Updated"}},"type":"object","title":"ConsentVault","description":"Per-customer consent record.\n\nEach boolean represents explicit opt-in (True) or opt-out (False).\nNew customers are provisioned with all channels enabled by default,\nmatching DPDP's \"consent given at sign-up\" assumption.  Operators\nshould set to False on unsubscribe events.","example":{"email":true,"last_updated":"2024-11-01T10:30:00+00:00","push":true,"whatsapp":false}},"CreativeRequest":{"properties":{"customer_id":{"type":"string","title":"Customer Id"},"action":{"type":"string","title":"Action"},"channel":{"type":"string","title":"Channel"},"properties":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Properties"}},"type":"object","required":["customer_id","action","channel"],"title":"CreativeRequest"},"CustomerProfileResponse":{"properties":{"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"device_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Device Type"},"active_hours":{"items":{"type":"integer"},"type":"array","title":"Active Hours"},"preferred_categories":{"items":{"type":"string"},"type":"array","title":"Preferred Categories"},"customer_id":{"type":"string","title":"Customer Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["customer_id","created_at"],"title":"CustomerProfileResponse"},"EventIdentifiers":{"properties":{"device_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Device Id"},"email_hash":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email Hash"},"phone_hash":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Hash"},"cookie_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cookie Id"},"loyalty_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Loyalty Id"}},"type":"object","title":"EventIdentifiers"},"EventIngestionResponse":{"properties":{"status":{"type":"string","title":"Status","default":"processed"},"event_id":{"type":"string","title":"Event Id"},"customer_id":{"type":"string","title":"Customer Id"},"updated_twin":{"$ref":"#/components/schemas/TwinStateResponse"},"nba_decision":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Nba Decision"},"final_action":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Final Action"},"final_channel":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Final Channel"},"consent_gate_applied":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Consent Gate Applied"},"generated_message":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Generated Message"},"identity_resolution":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Identity Resolution"}},"additionalProperties":true,"type":"object","required":["event_id","customer_id","updated_twin"],"title":"EventIngestionResponse"},"EventPayload":{"properties":{"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"},"event_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Event Type"},"metadata":{"additionalProperties":true,"type":"object","title":"Metadata"},"properties":{"additionalProperties":true,"type":"object","title":"Properties"},"source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source"}},"additionalProperties":true,"type":"object","title":"EventPayload"},"EventResponse":{"properties":{"event_id":{"type":"string","title":"Event Id"},"event_type":{"type":"string","title":"Event Type"},"source":{"type":"string","title":"Source"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"identifiers":{"$ref":"#/components/schemas/EventIdentifiers"},"customer_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Customer Id"},"properties":{"additionalProperties":true,"type":"object","title":"Properties"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["event_id","event_type","source","timestamp","identifiers","properties","created_at"],"title":"EventResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"NBAResponse":{"properties":{"customer_id":{"type":"string","title":"Customer Id"},"action":{"type":"string","title":"Action"},"confidence":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Confidence"},"reason":{"type":"string","title":"Reason"},"consent_gate_applied":{"type":"boolean","title":"Consent Gate Applied"},"final_channel":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Final Channel"}},"type":"object","required":["customer_id","action","confidence","reason","consent_gate_applied"],"title":"NBAResponse","example":{"action":"send_coupon","confidence":0.91,"consent_gate_applied":false,"customer_id":"cust_abc123","final_channel":"email","reason":"Customer abandoned cart with intent score 87 and fatigue score 12"}},"TwinStateResponse":{"properties":{"customer_id":{"type":"string","title":"Customer Id"},"journey_stage":{"type":"string","title":"Journey Stage"},"intent_score":{"type":"integer","title":"Intent Score"},"churn_risk":{"type":"integer","title":"Churn Risk"},"fatigue_score":{"type":"integer","title":"Fatigue Score"},"conversion_probability":{"type":"integer","title":"Conversion Probability"},"segment":{"type":"string","title":"Segment"},"preferred_channel":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Preferred Channel"},"next_best_action":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Next Best Action"},"raw_counters":{"additionalProperties":true,"type":"object","title":"Raw Counters"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["customer_id","journey_stage","intent_score","churn_risk","fatigue_score","conversion_probability","segment","updated_at"],"title":"TwinStateResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}